+++ example/android/app/src/main/kotlin/br/com/example/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package br.com.example.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}