+++ lib/src/disposeble.dart
@@ -1,0 +1,5 @@
+part of './dependencies.dart';
+
+abstract class Disposeble {
+  void dispose();
+}