+++ lib/src/exception.dart
@@ -2,9 +2,7 @@ part of './dependencies.dart';
 
 class DependencyManagerError implements Exception {
   final String message;
-
   const DependencyManagerError(this.message);
-
   @override
   String toString() => message;
 }