+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-## 1.0.0 primeiro release
+## 1.0.0
+* primeiro release
 
-## 1.0.1 Implementação do objeto DependencyManagerError
+## 1.0.1
+* Implementação do objeto DependencyManagerError
 
-## 1.0.2 Ajuste no método Dependencies.get
+## 1.0.2
+* Ajuste no método Dependencies.get
 
-## 1.0.3 Implementação da função [Dependencies.replaceInstanceSingleton]
+## 1.0.3
+* Implementação da função [Dependencies.replaceInstanceSingleton]
 
-## 1.0.4 Implementação de tratamento erro[NoSuchMethodError] na função [dependencyDispose] 
+## 1.0.4
+* Implementação de tratamento erro[NoSuchMethodError] na função [dependencyDispose] 
+
+## 1.0.5 
+* Restruturação do package para que seja possível instalar em projetos no dart ou flutter
+* Implementação do objeto [Disposeble]
+* Remoção do objeto [DependencyState]